LayoutTests/http/tests/security/contentSecurityPolicy/frame-src-cross-origin-load-expected.txt

CONSOLE MESSAGE: Refused to load https://localhost:8443/security/contentSecurityPolicy/resources/alert-fail.html because it does not appear in the frame-src directive of the Content Security Policy.
ALERT: PASS
ALERT: PASS
IFrames blocked by CSP should generate a 'load' event, regardless of blocked state. This means they appear to be normal cross-origin loads, thereby not leaking URL information directly to JS.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS IFrame #1 generated a 'load' event.
PASS IFrame #2 generated a 'load' event.
PASS IFrame #3 generated a 'load' event.
PASS successfullyParsed is true

TEST COMPLETE
